Move convenience function to correct package.
